move main.rs to examples folder
